Self-signing works!
The bug was that SHA512 was misreporting its output size, and we had too many ff's in the padded signature input.
Got a key to be accepted by gpg import, which is good.  To debug GPG MPI's it's useful to use the `gcry_mpi_dump`
routine, which will dump an MPI in hex to the console.