Address keybase/keybase-issues#921.
I didn't track it all the way down, but there's an issue with browserify-zlib decrypting
our fake ZIP archives.  When the "flush" is sent with 0 in bytes, it returns a Z_BUF_ERROR,
as if it still wants more data.  I think it's safe to ignore this error, but we should recheck
this assumption.  If we turn out to be wrong, we might need to call into pako directly
as shown above.  Calling into pako directly has problems, though, since it will be included
in the node.js setting which will increase code bloat.