get the right generated time for PGP keys
We were passing @timestamp as the key generation time, which is obtained
from some bits at the start of the key packet. Apparently for keys that
have been around for a long time and had their lifetimes extended, this
can be an old value. Instead, use the generated time that we get from
get_expire_time(), which searches for the set of times that result in
the latest expiration.

See https://github.com/keybase/keybase-issues/issues/1686, where this
resulted in us mistakenly thinking that a valid key was expired.