Add `assert_pgp_hash` callback (#190)
Add `assert_pgp_hash` optional callback argument to unboxing / verifying
functions. It will be called every time a Signature is being verified, and
provides an interface to fail the verification. For example, callback could
check for insecure hashing functions and return an error that using said
function is forbidden.

Add a way to pass `hasher` to boxing / signing function. This is mostly used in
tests for the `assert_pgp_hash` feature.